2022-4-5 修复与更新
1. 新增了接口：post /admin/choice/{adminId}(选择管理员)
2. 新增了ws的channel：adminList(管理员列表)

+++ src/main/java/com/whl/messagesystem/service/admin/AdminServiceImpl.java
@@ -1,15 +1,27 @@
 package com.whl.messagesystem.service.admin;
 
+import com.alibaba.fastjson.JSONObject;
+import com.whl.messagesystem.commons.channel.Channel;
+import com.whl.messagesystem.commons.channel.management.AdminListChannel;
 import com.whl.messagesystem.commons.constant.ResultEnum;
 import com.whl.messagesystem.commons.constant.RoleConstant;
 import com.whl.messagesystem.commons.constant.StringConstant;
 import com.whl.messagesystem.commons.utils.ResultUtil;
+import com.whl.messagesystem.commons.utils.WsResultUtil;
 import com.whl.messagesystem.dao.AdminDao;
+import com.whl.messagesystem.dao.GroupDao;
+import com.whl.messagesystem.dao.PublicGroupDao;
+import com.whl.messagesystem.dao.UserAdminDao;
 import com.whl.messagesystem.model.Result;
 import com.whl.messagesystem.model.dto.AdminInfo;
 import com.whl.messagesystem.model.dto.AdminRegisterDTO;
 import com.whl.messagesystem.model.dto.SessionInfo;
 import com.whl.messagesystem.model.entity.Admin;
+import com.whl.messagesystem.model.entity.Group;
+import com.whl.messagesystem.model.entity.PublicGroup;
+import com.whl.messagesystem.model.entity.UserAdmin;
+import com.whl.messagesystem.service.group.GroupService;
+import com.whl.messagesystem.service.message.MessageServiceImpl;
 import com.whl.messagesystem.service.session.SessionService;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.ObjectUtils;
@@ -17,6 +29,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.socket.TextMessage;
 
 import javax.annotation.Resource;
 import javax.servlet.http.HttpSession;
@@ -37,7 +50,22 @@ public class AdminServiceImpl implements AdminService {
     AdminDao adminDao;
 
     @Resource
+    GroupDao groupDao;
+
+    @Resource
+    UserAdminDao userAdminDao;
+
+    @Resource
+    PublicGroupDao publicGroupDao;
+
+    @Resource
     SessionService sessionService;
+
+    @Resource
+    GroupService groupService;
+
+    @Resource
+    MessageServiceImpl messageService;
 
     @Override
     public ResponseEntity<Result> getAdminList() {
@@ -66,11 +94,17 @@ public class AdminServiceImpl implements AdminService {
 
             if (adminDao.getAdminCountByAdminName(adminName) > 0) {
                 log.error("管理员已存在");
-                return ResponseEntity.ok(new Result(ResultEnum.ERROR.getStatus(), "管理员已存在", null));
+                return ResponseEntity.ok(new Result<>(ResultEnum.ERROR.getStatus(), "管理员已存在", null));
             }
 
             if (adminDao.insertAnAdmin(admin)) {
-                log.info("管理员注册成功");
+                admin = adminDao.selectAdminByAdminName(adminName);
+
+                // 实时更新管理员列表
+                Channel channel = new AdminListChannel();
+                String message = JSONObject.toJSONString(WsResultUtil.registerAdmin(admin));
+                messageService.publish(channel.getChannelName(), new TextMessage(message));
+
                 return ResponseEntity.ok(ResultUtil.success());
             }
 
@@ -82,7 +116,7 @@ public class AdminServiceImpl implements AdminService {
     }
 
     @Override
-    public ResponseEntity<Result> deleteAdmin(String adminId, HttpSession session) {
+    public synchronized ResponseEntity<Result> deleteAdmin(String adminId, HttpSession session) {
         try {
             if (StringUtils.isEmpty(adminId)) {
                 throw new NullPointerException("参数为空");
@@ -90,10 +124,31 @@ public class AdminServiceImpl implements AdminService {
 
             SessionInfo sessionInfo = (SessionInfo) session.getAttribute(StringConstant.SESSION_INFO);
             if (RoleConstant.ADMIN.equals(sessionInfo.getRole()) && adminId.equals(sessionInfo.getAdmin().getAdminId())) {
+                // 只剩一个管理员时是不允许注销的
+                if (adminDao.selectAdminCount() == 1) {
+                    log.warn("当前只剩一个管理员，无法注销");
+                    return ResponseEntity.ok(ResultUtil.error("当前只剩一个管理员，无法注销"));
+                }
+
+                // 放弃管理自己的私有分组
+                List<Group> privateGroups = groupDao.selectGroupsByAdminId(Integer.parseInt(adminId));
+                privateGroups.stream().map(Group::getGroupId).forEach(groupId -> groupService.giveUpManagePrivateGroup(groupId, session));
+
+                // 解散自己管理的公共分组
+                List<PublicGroup> publicGroups = publicGroupDao.selectPublicGroupsWithAdminId(Integer.parseInt(adminId));
+                publicGroups.stream().map(PublicGroup::getGroupId).forEach(groupId -> groupService.dismissPublicGroup(groupId));
+
                 if (adminDao.deleteAdminByAdminId(Integer.parseInt(adminId))) {
                     sessionService.logout(session);
+
+                    // 实时更新管理员列表
+                    AdminListChannel adminListChannel = new AdminListChannel();
+                    String message = JSONObject.toJSONString(WsResultUtil.deleteAdmin(adminId));
+                    messageService.publish(adminListChannel.getChannelName(), new TextMessage(message));
+
                     return ResponseEntity.ok(ResultUtil.success());
                 }
+
                 throw new SQLException("admin表删除记录失败");
             }
 
@@ -122,12 +177,39 @@ public class AdminServiceImpl implements AdminService {
                 sessionInfo.setAdmin(admin);
                 session.setAttribute(SESSION_INFO, sessionInfo);
 
+                AdminListChannel adminListChannel = new AdminListChannel();
+                String message = JSONObject.toJSONString(WsResultUtil.updateAdminNameAndPassword(admin));
+                messageService.publish(adminListChannel.getChannelName(), new TextMessage(message));
+
                 return ResponseEntity.ok(ResultUtil.success());
             }
 
             throw new SQLException("admin表更新记录失败");
         } catch (Exception e) {
             log.error("更新管理员的用户名与密码失败，参数：{}，异常信息：{}", adminInfo, e.getMessage());
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(ResultUtil.error());
+        }
+    }
+
+    @Override
+    public ResponseEntity<Result> choiceAnAdmin(String adminId, HttpSession session) {
+        try {
+            if (StringUtils.isEmpty(adminId)) {
+                throw new NullPointerException("参数为空");
+            }
+
+            SessionInfo sessionInfo = (SessionInfo) session.getAttribute(SESSION_INFO);
+            String userId = sessionInfo.getUser().getUserId();
+
+            if (userAdminDao.insertAnUserAdmin(new UserAdmin(userId, adminId))) {
+                Admin admin = adminDao.selectAdminByAdminId(Integer.parseInt(adminId));
+                sessionInfo.setAdmin(admin);
+                return ResponseEntity.ok(ResultUtil.success());
+            }
+
+            throw new SQLException("user_admin表插入失败");
+        } catch (Exception e) {
+            log.error("选择管理员异常，参数：{}，异常信息：{}", adminId, e.getMessage());
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(ResultUtil.error());
         }
     }
